+++ use_cases/classification_exp/utils.py
@@ -1,5 +1,0 @@
-import os
-
-
-def get_script_dir():
-    return os.path.dirname(os.path.realpath(__file__))