+++ use_cases/question_answering/bhh_object_count/__init__.py
@@ -1,3 +1,0 @@
-from adalflow.utils import setup_env
-
-setup_env()