+++ use_cases/question_answering/bhh_object_count/debug_trainer.py
@@ -1,1 +1,0 @@
-"""Here we show how we can visualize the gradients and a single step optimizer values to ensure the optimizer works fine before training in larger scale"""